requirements.txt
python-telegram-bot==20.7
vk-api==11.9.9
pandas==2.1.4
openpyxl==3.1.2


